feat: integrate user-service with iam-service
- Add Feign client to call iam-service for user data
- Create internal user endpoint in iam-service
- Update UserProfileService to fetch setupRequired from iam-service
- Update JWT handling to match iam-service (Long userId, sub claim)
- Update UserProfile entity to use Long for userId
- Remove comment lines

+++ src/main/java/az/fitnest/userservice/security/JwtAuthenticationFilter.java
@@ -30,7 +30,7 @@ public class JwtAuthenticationFilter extends OncePerRequestFilter {
         if (header != null && header.startsWith("Bearer ")) {
             String token = header.substring(7);
             try {
-                Integer userId = jwtUtil.getUserIdFromToken(token);
+                Long userId = jwtUtil.getUserIdFromToken(token);
 
                 Authentication authentication =
                         new UsernamePasswordAuthenticationToken(userId, null, null);
