+++ src/main/java/az/fitnest/userservice/user/adapter/persistence/UserProfileRepository.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface UserProfileRepository extends JpaRepository<UserProfile, Integer>{
+public interface UserProfileRepository extends JpaRepository<UserProfile, Long>{
 	
     Optional<UserProfile> findByUserId(Long userId);
 }
